Orca PyTorch spark Estimator add evaluate (#2760)
* add spark pytorch evaluate & testcase

* move to_sample

* fix submodule number
